[FP-76] Refactor: 코드 리뷰 반영하여 리액트 쿼리 함수들 hooks에서 apis로 이동
