+++ src/types/activity.ts
@@ -32,3 +32,12 @@ export type HeaderProps = {
   reviewCount: number;
   address: string;
 };
+
+export interface MyInformation {
+  id: number;
+  email: string;
+  nickname: string;
+  profileImageUrl: null;
+  createdAt: string;
+  updatedAt: string;
+}
